Organize LibraryViewController into separate files
